[Feat] #78 - GroupAddViewController 500줄 초과로 인해 Extension 파일 분리
